requirements.txt
# Core dependencies
kiteconnect>=4.1.0
pandas>=1.3.0
streamlit>=1.28.0
plotly>=5.0.0

# Database
psycopg2-binary>=2.9.0

# Web scraping
requests>=2.26.0
beautifulsoup4>=4.9.0
selenium>=4.0.0

# Data processing
openpyxl>=3.0.0
yfinance>=0.1.70
nsepy>=0.8

# Telegram and scheduling
python-telegram-bot>=13.0
schedule>=1.1.0

# Environment and logging
python-dotenv>=0.19.0
coloredlogs>=15.0